requirements.txt
langchain
openai
requests
beautifulsoup4
newspaper3k
google-generativeai
python-dotenv
requests
beautifulsoup4
